multilib: fix wrong arch classification
This partly reverts e6acde1a9 commit where is no clear
justification why the change was actually done.  Because the
pg_config.h was not ready for ppc64p7 and aarch64 -- it was clear
breakage.  In 1242873 discussion, it appeared that ppc64p7 needs
special handling as it is compatible/equivalent with ppc64.

Resolves: rhbz#1242873
Version: 9.4.4-3