spec: drop the old %beta hack about %__os_install_post
This always was broken and the %__os_install_post macro was
redefined also for non-beta builds;  that's because
%{?x:%global y z} defines %y even when %x is zero.

If we really wanted this %beta/%post hack, it would have to look
similarly to:
%{lua:if rpm.expand("%beta") ~= "0" then rpm.define('...') end}
On the other hand, I'm not sure that this is actually needed
nowadays so I'm removing it completely.  Please let us know if you
happen to need the %__os_install_post redefinition.

Originally added by commit 9636c11b0d6049d23.

Version: 9.6.3-5